+++ backend/src/Response/CaptainFinancialSystem/CaptainFinancialSystemAccordingOnOrderResponse.php
@@ -8,9 +8,9 @@ class CaptainFinancialSystemAccordingOnOrderResponse
     
     public string $categoryName;
     
-    public int $countKilometersFrom;
+    public float $countKilometersFrom;
 
-    public int $countKilometersTo;
+    public float $countKilometersTo;
 
     public float $amount;
